Fix types for user events, they do not contain groups
