+++ workos/resources/events.py
@@ -1,8 +1,9 @@
 from typing import Generic, Literal, TypeVar, Union
 from typing_extensions import Annotated
 from pydantic import Field
-from workos.resources.directory_sync import DirectoryGroup, DirectoryUser
+from workos.resources.directory_sync import DirectoryGroup
 from workos.resources.workos_model import WorkOSModel
+from workos.types.directory_sync.directory_user import DirectoryUser
 from workos.types.events.directory_group_membership_payload import (
     DirectoryGroupMembershipPayload,
 )
@@ -121,6 +122,11 @@ Event = Annotated[
         DirectoryGroupCreatedEvent,
         DirectoryGroupDeletedEvent,
         DirectoryGroupUpdatedEvent,
+        DirectoryUserCreatedEvent,
+        DirectoryUserDeletedEvent,
+        DirectoryUserUpdatedEvent,
+        DirectoryUserAddedToGroupEvent,
+        DirectoryUserRemovedFromGroupEvent,
     ],
     Field(..., discriminator="event"),
 ]
